requirements.txt
streamlit
tensorflow-cpu
transformers
numpy
gdown